fix: resolve all code review issues (apr 2026)
Consumer.dequeueItem(): null check su item, stampa messaggio se buffer vuoto (BUG-1)
UnsynchronizedOrderBuffer.enqueue(): aggiunto totItem++ (BUG-2)
SequentialSimulation: Thread.currentThread().interrupt() nel catch (BUG-3)
Buffer: aggiunto getTotItem() all'interfaccia (ARCH-1)
Buffer.dequeue(): Javadoc sul comportamento null (ARCH-2)
Simulation.run() rinominato execute() (ARCH-3)
Menu.launch(): rimosso in.close() (ARCH-4)
OrderBuffer, UnsynchronizedOrderBuffer: queue dichiarato final (STYLE-1)
Item.weight: float -> double (STYLE-2)
Producer, Consumer: campo queue rinominato buffer (STYLE-3)

+++ ProducerConsumer/src/com/github/floshdev/producerconsumer/model/Item.java
@@ -3,7 +3,7 @@ package com.github.floshdev.producerconsumer.model;
 public class Item {
 
 	private final int idItem;
-	private final float weight;
+	private final double weight;
 	private final Coordinate origin;
 	private final Coordinate destination;
 	
@@ -18,7 +18,7 @@ public class Item {
 		return idItem;
 	}
 
-	public float getWeight() {
+	public double getWeight() {
 		return weight;
 	}
